requirements.txt
boto3
requests
python-dotenv
json
datetime
groq